refactor: split pull_request.py into a package
760 lines and twelve functions in one module, with _review_code alone running
176 lines. Split along the same seams as the existing handlers/comments/
package, which already proved the shape:

    classify.py  file kind, review ordering, blast radius -- pure, no I/O
    analysis.py  risk/title analysis + reviewer summary
    review.py    AI code review + Reviews-API posting
    gaps.py      test-coverage gap detection
    report.py    assembly of the sticky comment body

handle() stays in __init__.py and calls the sub-analyses as module-level names,
so one module still decides what runs, in what order, and whether anything is
posted. Largest file is now 227 lines.

Behaviour is unchanged. Every external import path still resolves --
handlers/comments/reviewer.py (_blast_radius), evals/run.py (_review_code),
server.py (handle) -- via __all__ re-exports.

Five test patch targets moved to the module that now owns the name. Two of
them were already no-ops before this change: tests patched
app.handlers.pull_request.validate_pr_analysis and .check_pr_title_update, but
_analyze_pr resolved those through its own namespace, so the patches never
applied and the assertions passed against the real implementations. They now
patch the module that actually owns each name.

Also replaces three f['filename'] subscripts with .get() -- the same KeyError
shape already fixed in summarizer.py -- and names the review budget constants
(MAX_ISSUES_PER_FILE, MAX_DIFF_CHARS, LOW_CONFIDENCE_THRESHOLD) that were
inline slices.
